O usuário já pode definir uma função

+++ PostHtml/src/workload/Sketch.java
@@ -5,28 +5,22 @@
  */
 package workload;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.Map;
-
 /**
  *
  * @author elmr
  */
 public abstract class Sketch extends Expr {
     
-
-    
     public static Sketch mkColor(String colorName){
         return new Color(colorName);
     }
     
     public static Sketch mkProperty(Expr height, Expr width, Sketch color){
-        return new Property((Numeric)height, (Numeric)width, ((Color)color).color);
+        return new Property(height, width, ((Color)color).color);
     }
     //USADO EM FIGURAS QUE NECESSITAM DE APENAS UM PARAMETRO (CIRCLE E SQUARE)
     public static Sketch mkProperty(Expr height, Sketch color){
-        return new Property((Numeric)height, ((Color)color).color);
+        return new Property(height, ((Color)color).color);
     }
     
     public static Sketch mkShape(String shapeName){
